requirements.txt
beautifulsoup4==4.6.0
redis==2.10.6
requests==2.18.4
schedule==0.5.0
sqlalchemy1.2.1
